Working till LocalStorage

+++ app.js
@@ -9,9 +9,8 @@ class Book {
 }
 // kreiranje klase UI
 class UI {
-    // kreiranje metode addBookTolist klase UI
-    addBookToList(book) {
-        const list = document.getElementById('book-list');
+    // Metoda za kreiranje reda
+    createBookRow(book) {
         const row = document.createElement('tr');
         row.innerHTML = `
             <td>${book.title}</td>
@@ -19,7 +18,18 @@ class UI {
             <td>${book.isbn}</td>
             <td><a href="" class="delete">X</a></td>
         `;
-        list.appendChild(row);
+        return row;
+    }
+    // Metoda za dodavanje knjige u listu
+    addBookToList(book) {
+        const list = document.getElementById('book-list');
+        const row = this.createBookRow(book);
+        list.appendChild(row)
+    }
+    isBookExists(isbn) {
+        const list = document.getElementById('book-list');
+        const rows = list.querySelectorAll('tr');
+        return Array.from(rows).some(row => row.cells[2].textContent === isbn);
     }
     // kreiranje metode showAlert klase UI
     showAlert(message, className) {
@@ -62,6 +72,8 @@ document.getElementById('book-form').addEventListener('submit', (e) => {
 
     if(title === '' || author === '' || isbn ==='') {
         ui.showAlert('Niste popunili sva polja', 'error');
+    } else if (ui.isBookExists(isbn)) {
+        ui.showAlert('Knjiga sa ovim ISBN-om vec postoji!', 'error');
     } else {
         ui.addBookToList(book);
         ui.showAlert('Uspesno ste dodali knjigu!', 'success');
@@ -72,20 +84,17 @@ document.getElementById('book-form').addEventListener('submit', (e) => {
 // Listener za delete
 document.getElementById('book-list').addEventListener('click', e => {
     e.preventDefault();
-    const ui = new UI();
-    ui.showAlert('Uspesno ste obrisali knjigu!', 'error')
-    ui.deleteBook(e.target)
+    if (e.target.className === 'delete') {
+        const ui = new UI();
+        ui.showAlert('Uspesno ste obrisali knjigu!', 'success')
+        ui.deleteBook(e.target);
+    }
+    
 })
 
 
 
-
-// Probamo da sredimo da ne moze dve iste knjige.
-// Neki local storage da sacuva knjige.
-
-/*  - Refaktorisanje: Razmislite o izdvajanju logike za kreiranje novog reda (tr) i dodavanje u DOM u zasebnu funkciju,
-    unutar klase UI za veću čitljivost i ponovnu upotrebu.
-    
+/*  
     - Poboljšanje korisničkog interfejsa:
         Razmislite o dodavanju funkcionalnosti za uređivanje postojećih unosa knjiga.
         Poboljšajte stilizaciju alert divova; na primer, da poruka bude bliža korisniku kada se pojavi, može se dodati tranzicija ili animacija.
